Added launch files, hsv pose and zed pose remapped for prequal bt
