+++ src/auv_msgs/CMakeLists.txt
@@ -9,6 +9,7 @@ find_package(geometry_msgs REQUIRED)
 rosidl_generate_interfaces(${PROJECT_NAME}
   "msg/Detection.msg"
   "msg/DetectionArray.msg"
+  "msg/ControlCommand.msg"
   DEPENDENCIES std_msgs geometry_msgs
 )
