--- src/custom_interfaces/msg/ControlCommand.msg
@@ -1,4 +1,0 @@
-float32 delta_theta
-float32 delta_distance
-float32 target_depth
-uint8 stop_thrusters
