+++ src/pico_UART/pico_UART/uart_ros_bridge_3.py
@@ -2,7 +2,7 @@
 import rclpy
 from rclpy.node import Node
 from std_msgs.msg import Float32
-from custom_interfaces.msg import ControlCommand
+from auv_msgs.msg import ControlCommand
 from ctypes import *
 import serial
