+++ src/prequalification_bt/package.xml
@@ -14,7 +14,7 @@
   <depend>sensor_msgs</depend>
   <depend>std_msgs</depend>
   <depend>vision_msgs</depend>
-  <depend>custom_interfaces</depend>
+  <depend>auv_msgs</depend>
   <depend>tf2</depend>
   <depend>ament_index_cpp</depend>
   <depend>diagnostic_msgs</depend>
