+++ src/prequalification_bt/prequalification.xml
@@ -16,7 +16,7 @@ NOTE: If you change this file, you MUST RESTART the Behavior Tree node to reload
 
             <!-- 1. INITIALIZE -->
             <!-- NOTE: Change target_depth here (meters, positive-down) for the initial dive -->
-            <DiveToDepth target_depth="0.5" staystill="2.0"/>
+            <DiveToDepth target_depth="0.3" staystill="2.0"/>
 
             <!-- 2. GATE MISSION -->
             <Fallback name="find_gate">
