+++ src/ros_controls/config/front_cam.yaml
@@ -18,16 +18,18 @@
             min_depth: 0.01 # Min: 0.01, Max: 3.0
             max_depth: 15.0 # Min: 0.5, Max: 20.0
         
+        pos_tracking:
+        	pos_tracking_coord_sys: 'RIGHT_HANDED_Z_UP_X_FORWARD'
         object_detection:
             od_enabled: true
             detection_model: 'CUSTOM_YOLOLIKE_BOX_OBJECTS'
-            custom_onnx_file: '/home/cupcake/Prequal/src/ros_controls/ros_controls/best.onnx'
+            custom_onnx_file: '/home/cupcake/Desktop/prequal-task/src/ros_controls/models/Pq_gate.onnx'
             confidence_threshold: 50.0
             max_range: 15.0
 
             #custom_onnx_file: '/home/cupcake/Prequal/src/ros_controls/ros_controls/best.onnx'
             custom_onnx_input_size: 608 # Change to 640 if trained at 640
-            custom_obj_det_config_file: '/home/cupcake/ros2_ws/src/auv_bringup/config/custom_yolo.yaml'
+            custom_obj_det_config_file: '/home/cupcake/Desktop/prequal-task/src/ros_controls/config/custom_yolo.yaml'
       
             custom_class_count: 1
