+++ src/ros_controls/ros_controls/teleop.py
@@ -1,177 +1,70 @@
 #!/usr/bin/env python3
 import rclpy
-import sys, select, tty, termios, signal, time
+import sys, select, tty, termios, signal, math
 from rclpy.node import Node
-from std_msgs.msg import Float64
 from nav_msgs.msg import Odometry
-
-
-def clamp(v, lo, hi):
-    return max(lo, min(v, hi))
+from auv_msgs.msg import ControlCommand
 
 
 def is_data():
     return select.select([sys.stdin], [], [], 0) == ([sys.stdin], [], [])
 
 
-# ================= PID Controller =================
-class PIDController:
-    def __init__(self, kp, ki, kd, output_min, output_max, integral_limit=200.0):
-        self.kp = kp
-        self.ki = ki
-        self.kd = kd
-        self.output_min = output_min
-        self.output_max = output_max
-        self.integral_limit = integral_limit
-
-        self._integral = 0.0
-        self._prev_error = 0.0
-        self._prev_time = None
-
-    def compute(self, setpoint, measurement):
-        now = time.time()
-        if self._prev_time is None:
-            self._prev_time = now
-            return 0.0
-
-        dt = now - self._prev_time
-        if dt <= 0.0:
-            return 0.0
-        self._prev_time = now
-
-        error = setpoint - measurement
-        self._integral = clamp(
-            self._integral + error * dt,
-            -self.integral_limit, self.integral_limit
-        )
-        derivative = (error - self._prev_error) / dt
-        self._prev_error = error
-
-        output = (self.kp * error) + (self.ki * self._integral) + (self.kd * derivative)
-        return clamp(output, self.output_min, self.output_max)
-
-    def reset(self):
-        self._integral = 0.0
-        self._prev_error = 0.0
-        self._prev_time = None
-
-
 class TeleopNode(Node):
     def __init__(self):
-        super().__init__("teleop_thrusters")
+        super().__init__('teleop_node')
 
-        # ================= Controller Subscriptions =================
-        self.sub_front = self.create_subscription(
-            Float64, 'new_thrust_front', self.cb_front, 10)
-        self.sub_left = self.create_subscription(
-            Float64, 'new_thrust_left', self.cb_left, 10)
-        self.sub_right = self.create_subscription(
-            Float64, 'new_thrust_right', self.cb_right, 10)
+        self.cmd_pub = self.create_publisher(
+            ControlCommand, '/control_cmd', 10)
 
-        # ================= Odometry for depth feedback =================
         self.sub_odom = self.create_subscription(
             Odometry, '/odom', self.cb_odom, 10)
-        self.current_depth = 0.0
-        self.odom_received = False
 
-        # ================= Thruster Publishers =================
-        self.pubs = {
-            'back_propeller':    self.create_publisher(Float64, '/hydrogen/back_propeller/cmd_thrust', 10),
-            'right_propeller_1': self.create_publisher(Float64, '/hydrogen/right_propeller_1/cmd_thrust', 10),
-            'right_propeller_2': self.create_publisher(Float64, '/hydrogen/right_propeller_2/cmd_thrust', 10),
-            'left_propeller_1':  self.create_publisher(Float64, '/hydrogen/left_propeller_1/cmd_thrust', 10),
-            'left_propeller_2':  self.create_publisher(Float64, '/hydrogen/left_propeller_2/cmd_thrust', 10),
-        }
+        self.current_depth  = 0.0
+        self.target_depth   = 0.0
+        self.current_yaw    = 0.0
+        self.odom_received  = False
 
-        # ================= Controller Values =================
-        self.ctrl_values = {
-            'back_propeller':    0.0,
-            'left_propeller_2':  0.0,
-            'right_propeller_2': 0.0,
-        }
+        self.depth_step    = 0.1    # metres per keypress
+        self.distance_step = 0.1    # metres per keypress
+        self.yaw_step      = 0.0872665    # radians per keypress
 
-        self.manual_offsets = {k: 0.0 for k in self.pubs.keys()}
+        self.delta_theta    = 0.0
+        self.delta_distance = 0.0
+        self.stop_thrusters = False
 
-        self.step = 50.0
-        self.max_thrust = 50.0
-
-        # ================= Depth PID ? always active once odom arrives =================
-        self.depth_pid = PIDController(
-            kp=20.0,
-            ki=2.0,
-            kd=5.0,
-            output_min=-100.0,
-            output_max=100.0,
-            integral_limit=150.0
-        )
-        self.target_depth = 0.0
-        self.depth_step   = 0.1       # metres per keypress
-
-        # While A/D held: manual override replaces PID output
-        # When released: PID snaps setpoint to current depth and resumes
-        self.manual_vertical = 0.0    # non-zero only while A/D pressed
-        self.vertical_step   = 5.0
-        self.vertical_max    = 100.0
-
-        self.timer = self.create_timer(0.02, self.publish_all)
-
-    # ================= Callbacks =================
-    def cb_front(self, msg):
-        self.ctrl_values['back_propeller'] = msg.data
-
-    def cb_left(self, msg):
-        self.ctrl_values['left_propeller_2'] = msg.data
-
-    def cb_right(self, msg):
-        self.ctrl_values['right_propeller_2'] = msg.data
+        self.create_timer(0.02, self.publish_cmd)
 
     def cb_odom(self, msg):
         self.current_depth = -msg.pose.pose.position.z
+
+        # Extract yaw from quaternion
+        q = msg.pose.pose.orientation
+        siny = 2.0 * (q.w * q.z + q.x * q.y)
+        cosy = 1.0 - 2.0 * (q.y * q.y + q.z * q.z)
+        self.current_yaw = math.degrees(math.atan2(siny, cosy))
+
         if not self.odom_received:
-            # First odom ? initialise hold at current depth
             self.target_depth = self.current_depth
             self.odom_received = True
-            self.get_logger().info(f"Odom received. Holding depth {self.target_depth:.2f} m")
+            self.get_logger().info(
+                f"Odom received. Holding depth {self.target_depth:.2f} m")
 
-    # ================= Publishing =================
-    def publish_all(self):
-        if not self.odom_received:
-            # No odom yet ? publish zeros and wait
-            for pub in self.pubs.values():
-                msg = Float64()
-                msg.data = 0.0
-                pub.publish(msg)
-            return
+    def publish_cmd(self):
+        msg = ControlCommand()
+        msg.delta_theta    = self.delta_theta
+        msg.delta_distance = self.delta_distance
+        msg.target_depth   = self.target_depth
+        msg.stop_thrusters = self.stop_thrusters
 
-        if self.manual_vertical != 0.0:
-            # A/D pressed: bypass PID, drive manually
-            vert = self.manual_vertical
-        else:
-            # PID holds target depth automatically
-            vert = self.depth_pid.compute(self.target_depth, self.current_depth)
+        self.cmd_pub.publish(msg)
 
-        for name, pub in self.pubs.items():
-            ctrl = self.ctrl_values.get(name, 0.0)
+        # Reseting the stop thruster value to prevent stop_thrusters being published every iteration after space is pressed
+        self.stop_thrusters = False
 
-            if name in ('left_propeller_2', 'right_propeller_2'):
-                blended = ctrl + self.manual_offsets[name] + vert
-            elif name == 'back_propeller':
-                blended = ctrl + self.manual_offsets[name] + (vert * 0.9)
-            else:
-                blended = ctrl + self.manual_offsets[name]
-
-            msg = Float64()
-            msg.data = float(blended)
-            pub.publish(msg)
-
-    def stop_all(self):
-        for k in self.manual_offsets:
-            self.manual_offsets[k] = 0.0
-        self.manual_vertical = 0.0
-        # Snap hold to current depth and reset PID
-        self.target_depth = self.current_depth
-        self.depth_pid.reset()
-        self.publish_all()
+        # Reset per-cycle deltas after publishing
+        self.delta_theta    = 0.0
+        #self.delta_distance = 0.0
 
 
 def main(args=None):
@@ -179,20 +72,13 @@ def main(args=None):
     node = TeleopNode()
 
     print("""
-Teleop + Automatic Depth Hold
+Teleop sends ControlCommand to /control_cmd
 
-Depth hold is ALWAYS active. The AUV holds its current depth
-automatically. Use A/D to ascend/descend ? depth hold resumes
-at the new depth the moment you release.
-
-  A               : Ascend  (manual override while held)
-  D               : Descend (manual override while held)
-  Arrow Left/Right: Yaw
-  W / S           : Forward / Backward
-  I / K           : Pitch
-  L / R           : Roll left / right
-  SPACE           : Stop all + re-lock depth at current position
-  X               : Exit
+W / S        : Forward / Backward
+A / D        : Yaw left / right
+Up / Down    : Ascend / Descend
+SPACE        : Stop thrusters
+X            : Exit
 """)
 
     fd = sys.stdin.fileno()
@@ -200,7 +86,6 @@ at the new depth the moment you release.
     tty.setcbreak(fd)
 
     def exit_clean(*_):
-        node.stop_all()
         termios.tcsetattr(fd, termios.TCSADRAIN, old_settings)
         rclpy.shutdown()
         sys.exit(0)
@@ -209,82 +94,49 @@ at the new depth the moment you release.
 
     try:
         while rclpy.ok():
-            # Reset horizontal offsets every cycle (they are instantaneous)
-            for k in node.manual_offsets:
-                node.manual_offsets[k] = 0.0
-
-            # Reset manual vertical every cycle ? A/D must be held to keep override
-            prev_manual_vertical = node.manual_vertical
-            node.manual_vertical = 0.0
-
             if is_data():
                 ch = sys.stdin.read(1)
 
                 if ch == '\x1b':
                     seq = sys.stdin.read(2)
-
-                    if seq == '[C':    # yaw right
-                        node.manual_offsets['left_propeller_1']  =  node.step
-                        node.manual_offsets['right_propeller_1'] = -node.step
-
-                    elif seq == '[D':  # yaw left
-                        node.manual_offsets['left_propeller_1']  = -node.step
-                        node.manual_offsets['right_propeller_1'] =  node.step
+                    if seq == '[A':    # Up -> ascend
+                        node.target_depth -= node.depth_step
+                        print(f"  Target depth: {node.target_depth:.2f} m")
+                    elif seq == '[B':  # Down -> descend
+                        node.target_depth += node.depth_step
+                        print(f"  Target depth: {node.target_depth:.2f} m")
 
                 else:
                     ch = ch.lower()
-
                     if ch == 'w':
-                        node.manual_offsets['left_propeller_1']  = -node.step
-                        node.manual_offsets['right_propeller_1'] = -node.step
-
+                        if node.delta_distance < 0:
+                            node.delta_distance = -node.delta_distance
+                        else:
+                            node.delta_distance += node.distance_step
+                        print(f"  Target distance : {node.delta_distance:.2f} m")
                     elif ch == 's':
-                        node.manual_offsets['left_propeller_1']  =  node.step
-                        node.manual_offsets['right_propeller_1'] =  node.step
-
-                    elif ch == 'a':   # ascend ? manual override
-                        node.manual_vertical = clamp(
-                            node.vertical_step,
-                            -node.vertical_max, node.vertical_max)
-
-                    elif ch == 'd':   # descend ? manual override
-                        node.manual_vertical = clamp(
-                            -node.vertical_step,
-                            -node.vertical_max, node.vertical_max)
-
-                    elif ch == 'i':
-                        node.manual_offsets['left_propeller_2']  =  node.step
-                        node.manual_offsets['right_propeller_2'] = -node.step
-
-                    elif ch == 'k':
-                        node.manual_offsets['left_propeller_2']  = -node.step
-                        node.manual_offsets['right_propeller_2'] =  node.step
-
-                    elif ch == 'l':
-                        node.manual_offsets['left_propeller_2']  = -node.step
-                        node.manual_offsets['right_propeller_2'] =  node.step
-
-                    elif ch == 'r':
-                        node.manual_offsets['left_propeller_2']  =  node.step
-                        node.manual_offsets['right_propeller_2'] = -node.step
-
+                        node.delta_distance -= node.distance_step
+                        print(f"  Target distance : {node.delta_distance:.2f} m")
+                    elif ch == 'a':
+                        node.delta_theta -=node.yaw_step
+                        print(f"  Turning left : {node.delta_theta:.2f} radians")
+                    elif ch == 'd':
+                        node.delta_theta += node.yaw_step
+                        print(f"  Turning right : {node.delta_theta:.2f} radians")
+                    elif ch== 'o':
+                        node.delta_distance=0
                     elif ch == ' ':
-                        node.stop_all()
-                        print(f"  Stopped. Holding depth {node.target_depth:.2f} m")
-
+                        node.stop_thrusters = True
+                        node.delta_distance = 0.0
+                        node.delta_theta = 0.0
+                        print("  Thrusters stopped")
                     elif ch == 'x':
                         exit_clean()
-
-            # A/D just released ? snap PID setpoint to current depth
-            if prev_manual_vertical != 0.0 and node.manual_vertical == 0.0:
-                node.target_depth = node.current_depth
-                node.depth_pid.reset()
-                print(f"  Depth hold re-locked at {node.target_depth:.2f} m")
 
             rclpy.spin_once(node, timeout_sec=0.02)
     finally:
         exit_clean()
 
 
-if __name__ == "__main__":
+if __name__ == '__main__':
     main()
