--- src/ros_controls/ros_controls/yolo_box_provider,py
@@ -1,95 +1,0 @@
-import rclpy
-from rclpy.node import Node
-from sensor_msgs.msg import Image
-from cv_bridge import CvBridge
-from ultralytics import YOLO
-
-# Import the ZED custom box messages
-from zed_msgs.msg import CustomBox, CustomBoxArray
-
-class YOLOBoxProviderNode(Node):
-    def __init__(self):
-        super().__init__('yolo_box_provider_node')
-        self.bridge = CvBridge()
-
-        # --- LOAD YOUR AI MODEL ---
-        # IMPORTANT: Change this string to the actual path of your .onnx file on the AUV
-        model_path = '/home/user/ros2_ws/src/auv_bringup/models/best.onnx'
-        self.model = YOLO(model_path)
-
-        # Publisher to feed 2D boxes into the ZED Wrapper (Same exact topic as the HSV node!)
-        self.zed_custom_box_pub = self.create_publisher(
-            CustomBoxArray, 
-            '/zed2i_front/zed_node/custom_object_detection/input_rect_list', 
-            10
-        )
-
-        # Subscription to the raw image
-        self.img_sub = self.create_subscription(
-            Image, 
-            '/zed2i_front/zed_node/rgb/color/rect/image', 
-            self.image_callback, 
-            10
-        )
-        self.get_logger().info("YOLO 2D Box Provider for preq_gate Ready!")
-
-    def image_callback(self, msg):
-        # 1. Convert ROS image to OpenCV
-        frame = self.bridge.imgmsg_to_cv2(msg, desired_encoding='bgr8')
-
-        # 2. Run the image through your YOLO ONNX model
-        # verbose=False stops it from spamming your terminal with print statements every frame
-        results = self.model(frame, verbose=False)
-
-        # 3. Prepare the ZED Custom Box Array message
-        zed_boxes = CustomBoxArray()
-        zed_boxes.header = msg.header
-
-        for r in results:
-            for box in r.boxes:
-                # Extract the bounding box coordinates [x_min, y_min, x_max, y_max]
-                coords = box.xyxy[0].cpu().numpy().astype(int)
-                x_min, y_min, x_max, y_max = coords[0], coords[1], coords[2], coords[3]
-                
-                # Extract the confidence score
-                confidence = float(box.conf[0].cpu().numpy())
-                
-                # We only want to send confident detections to the ZED SDK
-                if confidence > 0.50:
-                    zed_box = CustomBox()
-                    
-                    # You can pull the actual class name from the model, but hardcoding 
-                    # it ensures the Master Node always gets exactly what it expects.
-                    zed_box.label = "preq_gate" 
-                    zed_box.probability = confidence
-                    
-                    # ZED expects the 4 corners explicitly: [TopLeft, TopRight, BottomRight, BottomLeft]
-                    zed_box.bounding_box_2d = [
-                        [int(x_min), int(y_min)],
-                        [int(x_max), int(y_min)],
-                        [int(x_max), int(y_max)],
-                        [int(x_min), int(y_max)]
-                    ]
-                    
-                    # Tell ZED it is allowed to start tracking this object across frames
-                    zed_box.is_tracking_ready = True 
-                    zed_boxes.boxes.append(zed_box)
-
-        # 4. If we found a gate, publish the 2D coordinates to the ZED SDK
-        if len(zed_boxes.boxes) > 0:
-            self.zed_custom_box_pub.publish(zed_boxes)
-
-
-def main(args=None):
-    rclpy.init(args=args)
-    node = YOLOBoxProviderNode()
-    try:
-        rclpy.spin(node)
-    except KeyboardInterrupt:
-        pass
-    finally:
-        node.destroy_node()
-        rclpy.shutdown()
-
-if __name__ == '__main__':
-    main()
