+++ src/ros_controls/setup.py
@@ -18,7 +18,7 @@ setup(
         
         (os.path.join('share', package_name, 'config'), glob(os.path.join('config', '*.yaml'))),
         
-        (os.path.join('share', package_name, 'model'), glob(os.path.join('model', '*.onnx'))),
+        (os.path.join('share', package_name, 'models'), glob(os.path.join('models', '*.onnx'))),
         
     ],
     install_requires=['setuptools'],
@@ -34,18 +34,11 @@ setup(
     },
     entry_points={
         'console_scripts': [
-            'teleop2 =ros_controls.teleop2:main',
-            'detection =ros_controls.obj_detect:main',
-            'viewer =ros_controls.image_viewer:main',
-            'viewer_3=ros_controls.viewer_3:main',
-            'detection_3=ros_controls.obj_detect_3:main',
-            'simple_cam=ros_controls.simple_cam:main',
-            'pyz_detect=ros_controls.pyzed_detect:main',
-            'hsv_detect=ros_controls.hsv_detector:main',
-            'combined_detect=ros_controls.combined_2:main',
-            'combined_detect_gpu=ros_controls.combined_3:main',
-            'combined_detect_gpu_2=ros_controls.combined_4:main',
-            #'dataset_collector =ros_controls:dataset_collector:main',
+            'teleop =ros_controls.teleop:main',
+            'combined_detections=ros_controls.combined_detections:main',
+            'exp_detections=ros_controls.combined_detection_'
+            'dataset_collector_front =ros_controls.dataset_collector_front:main',
+            'dataset_collector_down =ros_controls.dataset_collector_down:main',
         ],
     },
 )
